backend/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
python-multipart==0.0.9

diffusers==0.37.0
transformers==5.3.0
tokenizers==0.22.2
accelerate==1.13.0
safetensors==0.7.0

pillow==12.1.1
requests==2.32.5

opencv-python-headless==4.13.0.92
numpy==2.4.3
timm>=0.9

git+https://github.com/ChaoningZhang/MobileSAM.git